chore: move configs to mainnet/testnet folders
